+++ include/calibration/calibration_data.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include "calibration/aprilgrid.h"
-#include "libcbdetect/config.h"
+#include "calibration/aprilgrid.hpp"
 #include "utils/common_types.h"
 #include <basalt/calibration/calibration.hpp>
 #include <basalt/serialization/headers_serialization.h>
@@ -16,16 +15,17 @@ namespace basalt {
     public:
         OpenCVParams(int width, int height, bool adaptiveThresh, bool normalizeImage,
                      bool filterQuads, bool fastCheck, bool enableSubpixRefine)
-            : width(width),
-              height(height),
-              adaptiveThresh(adaptiveThresh),
-              normalizeImage(normalizeImage),
-              filterQuads(filterQuads),
-              fastCheck(fastCheck),
-              enableSubpixRefine(enableSubpixRefine) {}
+                : width(width),
+                  height(height),
+                  adaptiveThresh(adaptiveThresh),
+                  normalizeImage(normalizeImage),
+                  filterQuads(filterQuads),
+                  fastCheck(fastCheck),
+                  enableSubpixRefine(enableSubpixRefine) {}
 
         int width;
         int height;
+        // Flags
         bool adaptiveThresh;
         bool normalizeImage;
         bool filterQuads;
@@ -44,30 +44,9 @@ namespace basalt {
         EIGEN_MAKE_ALIGNED_OPERATOR_NEW
         CalibCornerData() = default;
 
-        CalibCornerData(const cbdetect::Corner &checkerboard_corner, const std::vector<cbdetect::Board> &boards) {
-            for (const auto& board : boards) {
-                int idx = 0;
-                for(const auto & i : board.idx) {
-                    for(int j = 0; j < i.size(); ++j) {
-                        if(i[j] >= 0) { // if the corner is valid
-                            int original_index = i[j];
-                            this->corners.emplace_back(checkerboard_corner.p[original_index].x, checkerboard_corner.p[original_index].y);
-                            this->corner_ids.push_back(idx++); // keep the original index
-                            this->radii.push_back(checkerboard_corner.r[original_index]);
-                        }
-                    }
-                }
-            }
-            // Lose the other fields
-            // std::vector<cv::Point2d> v1;
-            //  std::vector<cv::Point2d> v2;
-            //  std::vector<cv::Point2d> v3;
-            //  std::vector<double> score;
-        }
-
         CalibCornerData(const std::vector<cv::Point2f> cv_corners) {
             int idx = 0;
-            for(const auto & i : cv_corners) {
+            for (const auto &i: cv_corners) {
                 this->corners.emplace_back(i.x, i.y);
                 this->corner_ids.push_back(idx++);
                 this->radii.push_back(1.0); // NEED CHANGE THIS!!
@@ -92,23 +71,23 @@ namespace basalt {
     };
 
     using CalibCornerMap = tbb::concurrent_unordered_map<TimeCamId, CalibCornerData,
-                                                         std::hash<TimeCamId>>;
+            std::hash<TimeCamId>>;
 
     using CalibInitPoseMap =
             tbb::concurrent_unordered_map<TimeCamId, CalibInitPoseData,
-                                          std::hash<TimeCamId>>;
+                    std::hash<TimeCamId>>;
 
     class CalibParams {
     public:
         virtual ~CalibParams() = default;
 
         /*
-     * process method will be called in the detectCorners loop for each ManagedImage.
-     * AprilGrid detection right now takes in a Managed image, but preferably we can make the tag detector take in
-     * cv::Mat instead, since we have preprocessed the cv::Mat.
-     * Checkboard corner detection takes in cv::Mat, so currently it's converting to a cv::Mat before passing the
-     * image to the find_corners function.
-     * */
+         * process method will be called in the detectCorners loop for each ManagedImage.
+         * AprilGrid detection right now takes in a Managed image, but preferably we can make the tag detector take in
+         * cv::Mat instead, since we have preprocessed the cv::Mat.
+         * Checkboard corner detection takes in cv::Mat, so currently it's converting to a cv::Mat before passing the
+         * image to the find_corners function.
+         * */
         virtual void
         process(basalt::ManagedImage<uint16_t> &img_raw, CalibCornerData &ccd_good, CalibCornerData &ccd_bad) = 0;
 
@@ -124,57 +103,44 @@ namespace basalt {
 
     class AprilGridParams : public CalibParams {
     public:
-        AprilGridParams(const std::shared_ptr<AprilGrid> &april_grid) : ad(april_grid->getTagCols() * april_grid->getTagRows(), april_grid->getTagFamily(), april_grid->getLowId()) {
+        AprilGridParams(const std::shared_ptr<AprilGrid> &april_grid) : ad(
+                april_grid->getTagCols() * april_grid->getTagRows(), april_grid->getTagFamily(),
+                april_grid->getLowId()) {
             this->april_grid = april_grid;
             targetType = "aprilgrid";
         }
 
         AprilGridParams() = delete;
 
-        std::shared_ptr<AprilGrid> getParams() {return april_grid;}
+        std::shared_ptr<AprilGrid> getParams() { return april_grid; }
 
-        void process(basalt::ManagedImage<uint16_t> &img_raw, CalibCornerData &ccd_good, CalibCornerData &ccd_bad) override;
+        void
+        process(basalt::ManagedImage<uint16_t> &img_raw, CalibCornerData &ccd_good, CalibCornerData &ccd_bad) override;
 
     private:
         std::shared_ptr<AprilGrid> april_grid;
         ApriltagDetector ad;
     };
 
-    class CBCheckerboardParams : public CalibParams {
-    public:
-        CBCheckerboardParams(const std::shared_ptr<cbdetect::Params> &cb_params) {
-            this->cb_params = cb_params;
-            targetType = "checkerboard_cb";
-        }
-
-        CBCheckerboardParams() = delete;
-
-        std::shared_ptr<cbdetect::Params> getParams() {return cb_params;}
-
-        void process(basalt::ManagedImage<uint16_t> &img_raw, CalibCornerData &ccd_good, CalibCornerData &ccd_bad) override;
-
-    protected:
-        std::shared_ptr<cbdetect::Params> cb_params;
-    };
-
     class OpenCVCheckerboardParams : public CalibParams {
     public:
-        OpenCVCheckerboardParams(OpenCVParams& params) : cv_params(params) {
+        OpenCVCheckerboardParams(OpenCVParams &params) : cv_params(params) {
             // Access prams to get flags
             this->flags += params.adaptiveThresh ? cv::CALIB_CB_ADAPTIVE_THRESH : 0;
             this->flags += params.filterQuads ? cv::CALIB_CB_FILTER_QUADS : 0;
             this->flags += params.normalizeImage ? cv::CALIB_CB_NORMALIZE_IMAGE : 0;
             this->flags += params.fastCheck ? cv::CALIB_CB_FAST_CHECK : 0;
             this->enableSubpixRefine = params.enableSubpixRefine;
-          
+
             targetType = "checkerboard_opencv";
         }
 
         OpenCVCheckerboardParams() = delete;
 
-        OpenCVParams getParams() {return cv_params;}
+        OpenCVParams getParams() { return cv_params; }
 
-        void process(basalt::ManagedImage<uint16_t> &img_raw, CalibCornerData &ccd_good, CalibCornerData &ccd_bad) override;
+        void
+        process(basalt::ManagedImage<uint16_t> &img_raw, CalibCornerData &ccd_good, CalibCornerData &ccd_bad) override;
 
     protected:
         OpenCVParams cv_params;
@@ -187,10 +153,5 @@ namespace cereal {
     template<class Archive>
     void serialize(Archive &ar, basalt::CalibCornerData &c) {
         ar(c.corners, c.corner_ids, c.radii, c.seq);
-    }
-
-    template<class Archive>
-    void serialize(Archive &ar, basalt::CalibInitPoseData &c) {
-        ar(c.T_a_c, c.num_inliers, c.reprojected_corners);
     }
 }// namespace cereal
