--- src/gui/imports.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "recorder/presets.hpp"
-
-#include <glad/glad.h> // Initialize with gladLoadGL()
-#include <GLFW/glfw3.h> // Will drag system OpenGL headers
-#include "imgui.h"
-#include "imgui_impl_glfw.h"
-#include "imgui_impl_opengl3.h"
-#include "misc/cpp/imgui_stdlib.h"
-#include "implot.h"
-#include "immvision.h"
-#include <opencv2/opencv.hpp>
-#include <opencv2/core.hpp>
-#include <opencv2/imgproc.hpp>
-#include <opencv2/imgcodecs.hpp>
