--- src/utils/colors.h
@@ -1,28 +1,0 @@
-#pragma once
-
-#include "imgui.h"
-#include "utils/utils.hpp"
-
-static const ImVec4 light_blue = from_rgba(0, 174, 239, 255, true); // Light blue color for selected elements such as play button glyph when paused
-static const ImVec4 regular_blue = from_rgba(0, 115, 200, 255, true); // Checkbox mark, slider grabber
-static const ImVec4 light_grey = from_rgba(0xc3, 0xd5, 0xe5, 0xff, true); // Text
-static const ImVec4 dark_window_background = from_rgba(9, 11, 13, 255);
-static const ImVec4 almost_white_bg = from_rgba(230, 230, 230, 255, true);
-static const ImVec4 black = from_rgba(0, 0, 0, 255, true);
-static const ImVec4 transparent = from_rgba(0, 0, 0, 0, true);
-static const ImVec4 white = from_rgba(0xff, 0xff, 0xff, 0xff, true);
-static const ImVec4 scrollbar_bg = from_rgba(14, 17, 20, 255);
-static const ImVec4 scrollbar_grab = from_rgba(54, 66, 67, 255);
-static const ImVec4 grey{ 0.5f,0.5f,0.5f,1.f };
-static const ImVec4 dark_grey = from_rgba(30, 30, 30, 255);
-static const ImVec4 sensor_header_light_blue = from_rgba(80, 99, 115, 0xff);
-static const ImVec4 sensor_bg = from_rgba(36, 44, 51, 0xff);
-static const ImVec4 redish = from_rgba(255, 46, 54, 255, true);
-static const ImVec4 dark_red = from_rgba(200, 46, 54, 255, true);
-static const ImVec4 button_color = from_rgba(62, 77, 89, 0xff);
-static const ImVec4 header_window_bg = from_rgba(36, 44, 54, 0xff);
-static const ImVec4 header_color = from_rgba(62, 77, 89, 255);
-static const ImVec4 title_color = from_rgba(27, 33, 38, 255);
-static const ImVec4 device_info_color = from_rgba(33, 40, 46, 255);
-static const ImVec4 yellow = from_rgba(229, 195, 101, 255, true);
-static const ImVec4 green = from_rgba(0x20, 0xe0, 0x20, 0xff, true);
